warn on case sensitive searches
Because of historical reasons RT's searches
marked as case sensitive in RT::SearchBuilder::Limit, but we
expect them to be case insensitive.

Mixing case sensitive and case insensitive searches by the same
column causes problems for some columns on Pg and Oracle. Especially
those that should be indexed as indexes are different depending on
search kind.

This code catches such cases and throws a warning.