pre upgrade code
As we deprecated Type and new code uses Name.
Upgrade procedure is executed using new code, we have
to set Name equal to Type earlier so old upgrade scripts
still get correct information.

It's not abstracted to far and very special cased. We had
no need in such solution for years, it's possible we wouldn't
need more for another decade.