indexes on Groups table
All queries on Groups table should involve Domain column.
If Domain is not filtered then something is wrong.

We decided to go with case insensitive searches even on
columns where case is "controlled" - it's easier to implement
and isn't slower with proper indexes, so indexes are wrapped
into LOWER() functions for case sensitive DBs.

Type is deprecated and Name should be used, so we index it after
Domain. This covers loading of user defined (Instantless) groups
that were not covered before by any index longer than (Domain).

Log of all queries from an active instance showed that we do queries
like:

    Name = A AND ((Domain = B AND Instance = C) OR (Domain = D))

For such queries (Domain, Name) index doesn't provide enough
selectivity. Index on (Instance) can not be used. After all the query
is very selective as Domain = D is part that selects system role group
(4 records top). With (Domain, Name) index DB spends a lot of time
switching between the index and the table to re-check Instance = C
condition. If we have (Domain, Name, Instance) index then, at least
mysql, can skip jumping and just use index to check all conditions
without looking into table data.

Test were performed with SQL from 4.0 instance, but in master (4.2 to be)
we widened use of 1 as Instance value for system groups. This means
that in above query Instance is defined for every ORed branch and more
DBs can use full length of the index.

Instance is very selective column and short (like id), so we just index
it for situations when Type/Name is not defined, but instance is.

Domain, Type, Instance is kept for backwards compatibility and will go
away with Type column.